added safeguard node
