+++ manipulation/packages/frida_motion_planning/frida_motion_planning/motion_planning_server.py
@@ -2,11 +2,11 @@
 
 import rclpy
 from rclpy.node import Node
-from rclpy.action import ActionServer, ActionClient
+from rclpy.action import ActionServer
 from rclpy.callback_groups import ReentrantCallbackGroup
 from geometry_msgs.msg import TwistStamped, PoseStamped
 from std_msgs.msg import Bool
-from std_srvs.srv import SetBool, Empty
+from std_srvs.srv import SetBool, Empty, Trigger
 from frida_interfaces.action import MoveToPose, MoveJoints
 from frida_constants.manipulation_constants import (
     GET_COLLISION_OBJECTS_SERVICE,
@@ -35,28 +35,18 @@ from frida_constants.manipulation_constants import (
     TOGGLE_SERVO_SERVICE,
     GRIPPER_SET_STATE_SERVICE,
     MIN_CONFIGURATION_DISTANCE_TRESHOLD,
-    XARM_ROBOT_STATES_TOPIC,
-    XARM_CLEAN_ERROR_SERVICE,
-    XARM_MOTION_ENABLE_SERVICE,
     ESTOP_TOPIC,
+    MANIPULATION_ENSURE_ARM_READY_SERVICE,
 )
 
 from frida_interfaces.msg import CollisionObject
 from frida_motion_planning.utils.MoveItPlanner import MoveItPlanner
 from frida_motion_planning.utils.MoveItServo import MoveItServo
-from frida_motion_planning.utils.service_utils import (
-    move_joint_positions as send_joint_goal,
-)
-
-
 from frida_motion_planning.utils.XArmServices import XArmServices
 
 from trajectory_msgs.msg import JointTrajectory
 from sensor_msgs.msg import JointState
-from xarm_msgs.msg import RobotMsg
-from xarm_msgs.srv import SetInt16, SetInt16ById, Call as XArmCall
 from rclpy.qos import QoSProfile
-from controller_manager_msgs.srv import SwitchController
 
 
 class MotionPlanningServer(Node):
@@ -79,8 +69,6 @@ class MotionPlanningServer(Node):
         self.planner.set_planning_time(0.5)
         self.planner.set_planner(PICK_PLANNER)
 
-        # xarm states
-        self._arm_state: RobotMsg | None = None
         self._in_estop = False
         self.current_mode = -1
 
@@ -198,47 +186,18 @@ class MotionPlanningServer(Node):
             callback_group=self.callback_group,
         )
 
-        self.switch_controller_client = self.create_client(
-            SwitchController,
-            "/controller_manager/switch_controller",
-            callback_group=self.callback_group,
-        )
-        self._move_joints_client = ActionClient(
-            self,
-            MoveJoints,
-            MOVE_JOINTS_ACTION_SERVER,
+        self._ensure_arm_ready_client = self.create_client(
+            Trigger,
+            MANIPULATION_ENSURE_ARM_READY_SERVICE,
             callback_group=self.callback_group,
         )
 
-        self._estop_pub = self.create_publisher(Bool, ESTOP_TOPIC, 10)
-
-        self._clear_octomap_client = self.create_client(
-            Empty,
-            "/clear_octomap",
-            callback_group=self.callback_group,
-        )
-
-        self._clean_error_client = self.create_client(
-            XArmCall,
-            XARM_CLEAN_ERROR_SERVICE,
-            callback_group=self.callback_group,
-        )
-        self._motion_enable_client = self.create_client(
-            SetInt16ById,
-            XARM_MOTION_ENABLE_SERVICE,
-            callback_group=self.callback_group,
-        )
         self.create_subscription(
-            RobotMsg,
-            XARM_ROBOT_STATES_TOPIC,
-            self._on_arm_state,
+            Bool,
+            ESTOP_TOPIC,
+            lambda msg: setattr(self, "_in_estop", msg.data),
             10,
             callback_group=self.callback_group,
-        )
-
-        self._recovering = False
-        self.create_timer(
-            2.0, self._try_estop_recovery, callback_group=self.callback_group
         )
 
         self.get_logger().info("Motion Planning Server has been started")
@@ -310,7 +269,7 @@ class MotionPlanningServer(Node):
                 )
                 # Step 2: Use the planner's execution function
                 response.success = self.planner.execute_plan(
-                    trajectory_msg, get_arm_state=lambda: self._arm_state
+                    trajectory_msg, is_estop_active=lambda: self._in_estop
                 )
             else:
                 self.get_logger().error("Failed to parse trajectory file.")
@@ -366,7 +325,7 @@ class MotionPlanningServer(Node):
 
     def move_to_pose(self, goal_handle, feedback):
         """Perform the pick operation."""
-        self._ensure_arm_ready()
+        self._call_ensure_arm_ready()
         pose = goal_handle.request.pose
         target_link = goal_handle.request.target_link
         tolerance_position = (
@@ -389,7 +348,7 @@ class MotionPlanningServer(Node):
         if was_plan_successful:
             self.execute_trajectory(trajectory_plan)
             was_execution_successful = self.planner.execute_plan(
-                trajectory_plan, get_arm_state=lambda: self._arm_state
+                trajectory_plan, is_estop_active=lambda: self._in_estop
             )
             return was_execution_successful
         else:
@@ -428,7 +387,7 @@ class MotionPlanningServer(Node):
             self.get_logger().error(str(e))
             return False
 
-        self._ensure_arm_ready()
+        self._call_ensure_arm_ready()
         self.get_logger().info("Planning joint goal...")
         was_plan_successful, trajectory_plan = self.planner.plan_joint_goal(
             joint_positions,
@@ -440,7 +399,7 @@ class MotionPlanningServer(Node):
         if was_plan_successful:
             self.execute_trajectory(trajectory_plan)
             was_execution_successful = self.planner.execute_plan(
-                trajectory_plan, get_arm_state=lambda: self._arm_state
+                trajectory_plan, is_estop_active=lambda: self._in_estop
             )
             if was_execution_successful:
                 self.get_logger().info("Trajectory executed successfully.")
@@ -719,177 +678,20 @@ class MotionPlanningServer(Node):
 
         return response
 
-    def _on_arm_state(self, msg: RobotMsg):
-        self._arm_state = msg
-        if (msg.state == 4 or msg.err != 0) and not self._in_estop:
-            self._in_estop = True
-            self.get_logger().warn(
-                f"E-stop ACTIVATED (state={msg.state}, err={msg.err}) — broadcasting abort"
-            )
-            self._estop_pub.publish(Bool(data=True))
-
-    def _try_estop_recovery(self):
-        if not self._in_estop or self._recovering:
+    def _call_ensure_arm_ready(self):
+        if not self._ensure_arm_ready_client.wait_for_service(timeout_sec=5.0):
+            self.get_logger().warn("ensure_arm_ready service not available, skipping")
             return
-        self._recovering = True
-        try:
-            self.get_logger().info("E-stop active — attempting arm recovery...")
-            self._ensure_arm_ready()
-            if (
-                self._arm_state
-                and self._arm_state.state not in (3, 4)
-                and self._arm_state.err == 0
-            ):
-                self._in_estop = False
-                self.get_logger().warn(
-                    "E-stop CLEARED — arm recovered, going to nav_pose"
-                )
-                self._estop_pub.publish(Bool(data=False))
-                send_joint_goal(
-                    move_joints_action_client=self._move_joints_client,
-                    named_position="table_stare",
-                    velocity=0.3,
-                )
-                self._call_svc(
-                    self._clear_octomap_client, Empty.Request(), 5.0, "clear_octomap"
-                )
-        finally:
-            self._recovering = False
-
-    def _call_svc(self, client, request, timeout, label):
-        """Call a ROS service with a bounded timeout. Returns True on success."""
-        elapsed = 0.0
-        while not client.service_is_ready() and elapsed < timeout:
-            self.get_clock().sleep_for(rclpy.duration.Duration(seconds=0.1))
-            elapsed += 0.1
-        if not client.service_is_ready():
-            self.get_logger().warn(f"{label} -> service not ready after {timeout}s")
-            return False
-        future = client.call_async(request)
-        elapsed = 0.0
-        while not future.done() and elapsed < timeout:
-            self.get_clock().sleep_for(rclpy.duration.Duration(seconds=0.1))
-            elapsed += 0.1
-        if not future.done():
-            self.get_logger().error(f"{label} -> timed out after {timeout}s")
-            return False
-        return True
-
-    def _switch_controller(self, deactivate: list, activate: list, label: str) -> bool:
-        """Send one SwitchController request. Returns True on success."""
-        req = SwitchController.Request()
-        req.deactivate_controllers = deactivate
-        req.activate_controllers = activate
-        req.strictness = SwitchController.Request.BEST_EFFORT
-        future = self.switch_controller_client.call_async(req)
+        future = self._ensure_arm_ready_client.call_async(Trigger.Request())
         start = self.get_clock().now()
         while not future.done():
-            if (self.get_clock().now() - start).nanoseconds > 5e9:
-                self.get_logger().error(f"{label} -> timeout")
-                return False
-        result = future.result()
-        if result and result.ok:
-            self.get_logger().info(f"{label} -> ok")
-            return True
-        self.get_logger().error(f"{label} -> failed")
-        return False
-
-    def _reactivate_controller(self):
-        """Deactivate then reactivate xarm6_traj_controller. Returns True on success."""
-        if not self.switch_controller_client.wait_for_service(timeout_sec=5.0):
-            self.get_logger().error(
-                "_reactivate_controller -> switch_controller not available"
-            )
-            return False
-        self._switch_controller(
-            ["xarm6_traj_controller"], [], "deactivate xarm6_traj_controller"
-        )
-        return self._switch_controller(
-            [], ["xarm6_traj_controller"], "activate xarm6_traj_controller"
-        )
-
-    def _motion_enable(self) -> bool:
-        """Re-enable all motors (required after e-stop before mode/state changes)."""
-        req = SetInt16ById.Request()
-        req.id = 8  # all joints
-        req.data = 1  # enable
-        return self._call_svc(self._motion_enable_client, req, 5.0, "motion_enable")
-
-    def _reinit_mode1(self):
-        """Re-enable motors and transition arm through mode 0 → mode 1."""
-        self._motion_enable()
-        req_en = SetInt16.Request()
-        req_en.data = 0
-        self._call_svc(self.planner.mode_client, SetInt16.Request(), 5.0, "set_mode(0)")
-        self._call_svc(self.planner.state_client, req_en, 5.0, "set_state(0)")
-        req1 = SetInt16.Request()
-        req1.data = 1
-        self._call_svc(self.planner.mode_client, req1, 5.0, "set_mode(1)")
-        self._call_svc(
-            self.planner.state_client, req_en, 5.0, "set_state(0) post-mode1"
-        )
-
-    def _ensure_arm_ready(self):
-        """Check arm state and auto-recover from known-bad states before planning."""
-        if self._arm_state is None:
-            self.get_logger().warn(
-                "ensure_arm_ready -> no robot_states received yet, skipping check"
-            )
-            return
-
-        s = self._arm_state
-        needs_reinit = s.err != 0 or s.mode != 1
-        needs_reenable = s.state in (3, 4)
-
-        if not (needs_reinit or needs_reenable):
-            return  # arm already healthy, nothing to do
-
-        if s.err != 0:
-            self.get_logger().warn(f"ensure_arm_ready -> err={s.err}, cleaning error")
-            self._call_svc(
-                self._clean_error_client, XArmCall.Request(), 5.0, "clean_error"
-            )
-
-        if needs_reinit:
-            self.get_logger().warn(
-                f"ensure_arm_ready -> mode={s.mode} (expected 1), reinit mode 1"
-            )
-            self._reinit_mode1()
-
-        if needs_reenable:
-            self.get_logger().warn(
-                f"ensure_arm_ready -> state={s.state} (paused/stopped), re-enabling motion"
-            )
-            self._motion_enable()
-            req = SetInt16.Request()
-            req.data = 0
-            self._call_svc(self.planner.state_client, req, 5.0, "set_state(0)")
-
-        self._reactivate_controller()
-        self._wait_for_arm_ready(timeout_sec=5.0)
-        # Clear stale octomap voxels: when the arm stops mid-motion the camera captures the arm itself as an obstacle
-        self._call_svc(
-            self._clear_octomap_client, Empty.Request(), 5.0, "clear_octomap"
-        )
-
-    def _wait_for_arm_ready(self, timeout_sec: float = 5.0):
-        """Spin-wait until arm reports state 1/2 with no error, or timeout."""
-        deadline = self.get_clock().now() + rclpy.duration.Duration(seconds=timeout_sec)
-        while self.get_clock().now() < deadline:
-            if (
-                self._arm_state is not None
-                and self._arm_state.state in (1, 2)
-                and self._arm_state.err == 0
-            ):
+            if (self.get_clock().now() - start).nanoseconds > 15e9:
+                self.get_logger().error("ensure_arm_ready timed out")
                 return
             self.get_clock().sleep_for(rclpy.duration.Duration(seconds=0.1))
-        s = self._arm_state.state if self._arm_state else "?"
-        self.get_logger().warn(
-            f"ensure_arm_ready -> arm did not reach ready state within {timeout_sec}s (state={s})"
-        )
 
     def reset_xarm_controller(self, request, response):
-        self._reactivate_controller()
+        self._call_ensure_arm_ready()
         return response
 
     def joint_states_callback(self, msg: JointState):
